pyproject.toml
[project]
name = "langgraph-ai-agent-service-examples"
version = "0.1.0"
description = "langgraph-ai-agent-service-examples 저장소용 AI 에이전트 실습 예제"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "AI Agent Book" }]

dependencies = [
    # LangChain 핵심 (2025.12 기준)
    "langchain>=0.3.25",
    "langchain-core>=0.3.61",
    "langchain-openai>=0.3.18",

    # LangGraph
    "langgraph>=0.4.8",
    "langgraph-supervisor>=0.0.12",

    # 설정 관리
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",

    # 웹 UI
    "streamlit>=1.45.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.14.8",
    "mypy>=1.19.0",
    "pytest>=9.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["shared"]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP"]
ignore = ["E501"]
